[svn] Implement xdigit_to_xchar more efficiently.

+++ src/utils.h
@@ -109,8 +109,8 @@ struct wget_timer *wtimer_allocate PARAMS ((void));
 struct wget_timer *wtimer_new PARAMS ((void));
 void wtimer_delete PARAMS ((struct wget_timer *));
 void wtimer_reset PARAMS ((struct wget_timer *));
-long wtimer_elapsed PARAMS ((struct wget_timer *));
-long wtimer_granularity PARAMS ((void));
+double wtimer_elapsed PARAMS ((struct wget_timer *));
+double wtimer_granularity PARAMS ((void));
 
 char *html_quote_string PARAMS ((const char *));
